refactor: Update reels and categories endpoints
