+++ trends_domain/src/commonMain/kotlin/net/thechance/mena/trends/domain/repository/CategoryRepository.kt
@@ -4,7 +4,6 @@ import net.thechance.mena.trends.domain.entity.Category
 
 interface CategoryRepository {
     suspend fun getAllCategories(): List<Category>
-    suspend fun isCategoriesAlreadySelectedByUser(): Boolean
     suspend fun initializeUserCategories(categoriesIds: List<String>)
     suspend fun updateUserCategories(
         originalSelectedIds: List<String>,
